Split down viewdock into modules
